requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.2.0
alabaster==0.7.16
alembic==1.16.2
anyio==4.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astroid==3.3.8
asttokens==3.0.0
async-lru==2.0.4
asyncssh==2.17.0
atomicwrites==1.4.0
attrs==24.3.0
autopep8==2.0.4
babel==2.16.0
bcrypt==4.3.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
black==24.10.0
bleach==6.2.0
Brotli==1.0.9
certifi==2025.7.14
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.8
cloudpickle==3.0.0
colorama==0.4.6
colorlog==6.9.0
comm==0.2.1
contourpy==1.3.2
cookiecutter==2.6.0
cryptography==44.0.1
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dieboldmariano==1.1.0
diff-match-patch==20200713
dill==0.3.8
docstring-to-markdown==0.11
docutils==0.21.2
executing==0.8.3
fastjsonschema==2.20.0
filelock==3.18.0
flake8==7.1.1
fonttools==4.58.0
frozenlist==1.5.0
fsspec==2025.3.2
greenlet==3.2.3
h11==0.14.0
httpcore==1.0.2
httpx==0.28.1
idna==3.7
imagesize==1.4.1
importlib_metadata==8.5.0
inflection==0.5.1
intervaltree==3.1.0
ipykernel==6.29.5
ipython==8.30.0
isort==6.0.1
jedi==0.19.2
jellyfish==1.1.3
Jinja2==3.1.6
joblib==1.5.0
json5==0.9.25
jsonschema==4.23.0
jsonschema-specifications==2023.7.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keyring==25.6.0
kiwisolver==1.4.8
lightning-utilities==0.14.3
Mako==1.3.10
markdown-it-py==2.2.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.0
missingno==0.5.2
mistune==3.1.2
more-itertools==10.3.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook_shim==0.2.4
numpy==2.2.5
numpydoc==1.7.0
optuna==4.4.0
overrides==7.4.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.0
parso==0.8.4
pathspec==0.10.3
patsy==1.0.1
pexpect==4.8.0
pickleshare==0.7.5
pillow==11.2.1
pip==25.1
platformdirs==4.3.7
pluggy==1.5.0
ply==3.11
prometheus_client==0.21.1
prompt-toolkit==3.0.43
propcache==0.3.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.12.1
pycparser==2.21
pydocstyle==6.3.0
pyflakes==3.2.0
PyGithub==2.4.0
Pygments==2.19.1
PyJWT==2.10.1
pylint==3.3.5
pylint-venv==3.0.3
pyls-spyder==0.4.0
PyNaCl==1.5.0
pyparsing==3.2.3
PyQt5==5.15.10
PyQt5_sip==12.13.0
PyQtWebEngine==5.15.6
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
python-lsp-black==2.0.0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.12.2
python-slugify==5.0.2
PyYAML==6.0.2
requests==2.32.3
roman-numerals-py==3.1.0
Rtree==1.0.1
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
setuptools==78.1.1
SQLAlchemy==2.0.41
stack-data==0.2.0
statsmodels==0.14.4
superqt==0.7.3
sympy==1.14.0
tabulate==0.9.0
terminado==0.17.1
text-unidecode==1.3
textdistance==4.6.3
threadpoolctl==3.6.0
three-merge==0.1.1
tinycss2==1.4.0
tomli==2.0.1
tomlkit==0.13.2
torch==2.7.0
torchmetrics==1.7.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.2
ujson==5.10.0
Unidecode==1.3.8
urllib3==2.3.0
watchdog==4.0.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.8.0
whatthepatch==1.0.2
wheel==0.45.1
win-inet-pton==1.1.0
wrapt==1.17.0
yapf==0.40.2
yarl==1.18.0
zipp==3.21.0
